+++ three.js/src/index-arjs.js
@@ -1,0 +1,20 @@
+import Anchor from './new-api/arjs-anchor';
+import HitTesting from './new-api/arjs-hittesting';
+import { AnchorDebugUI, SessionDebugUI } from './new-api/arjs-debugui';
+import Session from './new-api/arjs-session';
+import Utils from './new-api/arjs-utils';
+import Context from './threex/arjs-context';
+import Profile from './threex/arjs-profile';
+import Source from './threex/arjs-source';
+
+export {
+  Anchor,
+  HitTesting,
+  AnchorDebugUI,
+  SessionDebugUI,
+  Session,
+  Utils,
+  Context,
+  Profile,
+  Source
+};