+++ three.js/src/index-threex-nft.js
@@ -1,0 +1,21 @@
+import ArMarkerControls from './threex/arjs-markercontrols-nft'; // nft-specific
+import ArMarkerHelper from './threex/threex-armarkerhelper';
+import ArSmoothedControls from './threex/threex-arsmoothedcontrols';
+import ArToolkitContext from './threex/arjs-context-nft'; // nft-specific
+import ArToolkitProfile from './threex/arjs-profile';
+import ArToolkitSource from './threex/arjs-source';
+import ArMultiMarkerControls from './markers-area/arjs-markersareacontrols';
+import ArMultiMakersLearning from './markers-area/arjs-markersarealearning'; // typo is spread over codebase
+import ArMultiMarkerUtils from './markers-area/arjs-markersareautils';
+
+export {
+  ArMarkerControls,
+  ArMarkerHelper,
+  ArSmoothedControls,
+  ArToolkitContext,
+  ArToolkitProfile,
+  ArToolkitSource,
+  ArMultiMarkerControls,
+  ArMultiMakersLearning,
+  ArMultiMarkerUtils
+};