+++ three.js/src/threex/threex-arclickability.js
@@ -1,4 +1,4 @@
-var THREEx = THREEx || {}
+import * as THREE from 'three';
 
 // TODO this is useless - prefere arjs-HitTesting.js
 
@@ -7,18 +7,18 @@ var THREEx = THREEx || {}
  * - seems an easy light layer for clickable object
  * - up to
  */
-THREEx.ARClickability = function (sourceElement) {
+const ARClickability = function (sourceElement) {
     this._sourceElement = sourceElement
     // Create cameraPicking
     var fullWidth = parseInt(sourceElement.style.width)
     var fullHeight = parseInt(sourceElement.style.height)
     this._cameraPicking = new THREE.PerspectiveCamera(42, fullWidth / fullHeight, 0.1, 100);
 
-    console.warn('THREEx.ARClickability works only in modelViewMatrix')
+    console.warn('ARClickability works only in modelViewMatrix')
     console.warn('OBSOLETE OBSOLETE! instead use THREEx.HitTestingPlane')
 }
 
-THREEx.ARClickability.prototype.onResize = function () {
+ARClickability.prototype.onResize = function () {
     var sourceElement = this._sourceElement
     var cameraPicking = this._cameraPicking
 
@@ -28,7 +28,7 @@ THREEx.ARClickability.prototype.onResize = function () {
     cameraPicking.updateProjectionMatrix();
 }
 
-THREEx.ARClickability.prototype.computeIntersects = function (domEvent, objects) {
+ARClickability.prototype.computeIntersects = function (domEvent, objects) {
     var sourceElement = this._sourceElement
     var cameraPicking = this._cameraPicking
 
@@ -45,6 +45,8 @@ THREEx.ARClickability.prototype.computeIntersects = function (domEvent, objects)
     return intersects
 }
 
-THREEx.ARClickability.prototype.update = function () {
+ARClickability.prototype.update = function () {
 
 }
+
+export default ARClickability;