+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.4.0-alpha-rc1
+- ES6 feature into master
+- new threejs location based
+- new dispose methods in arjs-context.js and arjs-source.js
+- use of artoolkit5-js instead of old jsartoolkit5 as base code
+- other bug fixes see PR #400
 # 3.3.3
 
 - Fixed bug which prevents `minDistance` and `maxDistance` working with a simulated location